Added functionality to exit button & resized start menu
I resized the screen menu to display the entire background since it kept zooming too far in when you ran the game. I also added code on the start_screen script to add functionality to the exit button so it stops the game and closes the window when the user clicks it.